LayoutTests/fast/dom/Text/text-should-not-have-attributes.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="test"></div>
<script>

description('This tests that Text nodes do no have attributes nor hasAttributes.');

shouldBeFalse('Text.prototype.hasOwnProperty("hasAttributes")');

var text = document.createTextNode('text');
shouldBeFalse('"hasAttributes" in text');
shouldBeFalse('"attributes" in text');

</script>
<script src="../../js/resources/js-test-post.js"></script>
